apps/dashboard/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finsight-dashboard"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "dash>=2.16",
  "plotly>=5.20",
  "httpx>=0.27",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "structlog>=24.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
